refactored heredoc_loop into its own file
