+++ include/minishell.h
@@ -159,6 +159,7 @@ int				reset_stdfds(t_data *core);
 // int				close_fds(t_data *core, t_pipe_fds fds);
 int				save_stdfds(t_data *core);
 int				init_pipes(t_data *core);
+int				heredoc_loop(char *limiter, int fd, int iscommand);
 
 int				exec_builtin(t_data *core, char *cmdpath, char **args, int is_exit);
